Support showing help for command when parse fails
This feature was left as todo because I wasn't sure exactly how to do
it. Now I think I found a decent way to do it. `Commands/All` exports
`help` which could be any command, but in the standard distribution it's
just re-exported from the builtin help command in `Commands/Help`. TS
will require it to conform to a specific signature.

In this way, only `Commands/All` is required to exist, which was my goal.