Refactor database managers
Previously the database manager took an existing MongoDB connection. Now
the manager is responsible for making its own connection. This allows
for a variety of databases to be hooked in with an adapter-like
architecture. All of the adapters implement the abstract DBManager class,
and the default export of the Database.ts file is the chosen manager /
adapter.

Note: this commit includes various side effects such as removal of
unused imports.