Allow for nickname mentions for mention-by-prefix
Previously we would only allow for plain mentions (`<@xxxxxx>`). It
seems, though, that Discord uses nickname mentions (`<@!xxxxxx>`) when
the end user mentions another Discord user (e.g., the bot). In the
spirit of the seamless processing of both nickname and plain mentions
within the `user` Parser, we now accept both nickname and plain mentions
for use in mention-as-prefix operation.